requirements.txt
protobuf
google-cloud-language
httplib2
google-auth-httplib2
